% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bridget.R
\name{bridget}
\alias{bridget}
\title{A Bridget Function}
\usage{
bridget(replace_col)
}
\arguments{
\item{Do}{you love cats? Defaults to TRUE.}
}
\description{
This function allows you to express your love of cats.
}
\examples{
bridget()
}
\keyword{none}
