% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DanE.R
\name{DanE}
\alias{DanE}
\title{A DElder Function}
\usage{
DanE()
}
\arguments{
\item{Do}{you love cats? Defaults to TRUE.}
}
\description{
This function allows you to express your love of cats.
}
\examples{
DanE()
}
\keyword{none}
