% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DanL.R
\name{DanL}
\alias{DanL}
\title{A DanL Function (jefe)}
\usage{
DanL()
}
\arguments{
\item{Do}{you love cats? Defaults to TRUE.}
}
\description{
This function allows you to express your love of cats.
}
\examples{
DanL()
}
\keyword{none}
