% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joes_message.R
\name{joes_message}
\alias{joes_message}
\title{A Joe's Function}
\usage{
joes_message()
}
\arguments{
\item{Do}{you love cats? Defaults to TRUE.}
}
\description{
This function allows you to express your love of cats.
}
\examples{
joes_message()
}
\keyword{none}
