% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/john_message.R
\name{john_message}
\alias{john_message}
\title{A John's Function}
\usage{
john_message(color1 = "blue", color2 = "yellow", color3 = "red")
}
\arguments{
\item{Do}{you love cats? Defaults to TRUE.}
}
\description{
This function allows you to express your love of cats.
}
\examples{
john_message()
}
\keyword{none}
