% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logan_message.R
\name{logan_message}
\alias{logan_message}
\title{A Logan Function}
\usage{
logan_message(recipient = "Akika")
}
\arguments{
\item{Do}{you love cats? Defaults to TRUE.}
}
\description{
This function allows you to express your love of cats.
}
\examples{
logan_message()
}
\keyword{none}
