% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mariona.R
\name{mariona}
\alias{mariona}
\title{A Mariona Function}
\usage{
mariona()
}
\arguments{
\item{Do}{you love cats? Defaults to TRUE.}
}
\description{
This function allows you to express your love of cats.
}
\examples{
mariona()
}
\keyword{none}
